refactor: decompose EventCreation 1,270-line component into modular sub-components
Extracted from EventCreation.jsx:
- utils/eventFormValidation.js (pure validateForm function, testable)
- components/RestoreDraftModal.jsx (restore draft modal)
- components/GuidelinesSection.jsx (guidelines card)
- components/EventDurationSelector.jsx (single/multi-day radio)
- components/DateTimeFields.jsx (date/time inputs for both duration types)
- components/LocationFields.jsx (virtual link vs physical address fields)
- components/RegistrationDatesFields.jsx (registration start/end dates)
- components/TagsInput.jsx (tag add/remove with keyboard support)
- components/StatsSection.jsx (static marketing stat cards)

EventCreation.jsx reduced from 1,270 to 541 lines (-57%).

Fixes #5512
