+++ src/components/common/EventCreation/EventCreation.jsx
@@ -1,4 +1,3 @@
-import { ArrowRightIcon, CalendarIcon, UsersIcon, ClipboardDocumentListIcon, TagIcon, CheckCircleIcon } from "@heroicons/react/24/outline";
 import { useState, useEffect } from "react";
 import { motion } from "framer-motion";
 import { toast } from "react-toastify";
@@ -9,6 +8,14 @@ import TicketsStep from "./components/TicketsStep";
 import GeneralInfoStep from "./components/GeneralInfoStep";
 import { exportAttendeesToCSV } from "../../../utils/exportCsv";
 import PreviewStep from "./components/PreviewStep";
+import RestoreDraftModal from "./components/RestoreDraftModal";
+import GuidelinesSection from "./components/GuidelinesSection";
+import EventDurationSelector from "./components/EventDurationSelector";
+import DateTimeFields from "./components/DateTimeFields";
+import LocationFields from "./components/LocationFields";
+import RegistrationDatesFields from "./components/RegistrationDatesFields";
+import TagsInput from "./components/TagsInput";
+import StatsSection from "./components/StatsSection";
 import {
   DRAFT_KEY,
   CREATION_STEPS,
@@ -18,35 +25,16 @@ import {
   todayString,
 } from "../../../constants/eventDefaults";
 import { API_ENDPOINTS, apiUtils } from "../../../config/api";
-import {
-  Calendar,
-  MapPin,
-  Link2,
-  Users,
-  Globe,
-  CalendarPlus,
-  CalendarX,
-  Map,
-  Navigation,
-  Compass,
-  Plus,
-} from "lucide-react";
 import { useFormSubmit } from "../../../hooks/useFormSubmit";
-import {
-  parseTimeToMinutes,
-  validateCoordinates,
-} from "../../../utils/eventCreationUtils";
-
+import { validateCoordinates } from "../../../utils/eventCreationUtils";
+import { validateForm } from "../../../utils/eventFormValidation";
 
 const EventCreation = () => {
   const prefersReducedMotion = useReducedMotion();
-  
+
   const [currentStep, setCurrentStep] = useState(CREATION_STEPS.FORM);
 
   const { handleSubmit: submitEventForm, isSubmitting, error: submitError, success: submitSuccess } = useFormSubmit(async (eventData) => {
-    // Auth is handled by the HttpOnly session cookie — apiUtils sends it
-    // automatically via withCredentials. Never read tokens from sessionStorage;
-    // setToken was removed as part of the HttpOnly cookie migration.
     if (!API_ENDPOINTS.EVENTS.CREATE) {
       await new Promise((resolve) => setTimeout(resolve, 1000));
       return;
@@ -72,136 +60,12 @@ const EventCreation = () => {
       });
     }
   }, [submitSuccess]);
+
   const [formData, setFormData] = useState(initialFormData);
   const [errors, setErrors] = useState({});
   const [newTag, setNewTag] = useState("");
-  // Track whether draft has been loaded to avoid overwriting on initial mount
   const [isDraftLoaded, setIsDraftLoaded] = useState(false);
   const [showRestoreModal, setShowRestoreModal] = useState(false);
-
-  const validateForm = () => {
-    const newErrors = {};
-
-    if (!formData.title.trim()) {
-      newErrors.title = "Event title is required";
-    } else if (formData.title.length < 3 || formData.title.length > 200) {
-      newErrors.title = "Title must be between 3 and 200 characters";
-    }
-
-    if (!formData.description.trim()) newErrors.description = "Event description is required";
-    if (!formData.category) newErrors.category = "Please select a category";
-
-    if (formData.isMultiDay) {
-      if (!formData.startDate) newErrors.startDate = "Start date is required";
-      if (!formData.endDate) newErrors.endDate = "End date is required";
-
-      if (formData.startDate && formData.endDate) {
-        if (new Date(formData.endDate) < new Date(formData.startDate)) {
-          newErrors.endDate = "End date must be after start date";
-        }
-      }
-    } else {
-      if (!formData.date) newErrors.date = "Event date is required";
-    }
-
-    if (!formData.startTime) newErrors.startTime = "Start time is required";
-    if (!formData.endTime) newErrors.endTime = "End time is required";
-
-    if (!newErrors.startTime && !newErrors.endTime && !formData.isMultiDay) {
-      // Convert time strings (HH:MM format) to minutes for proper comparison
-      const startMinutes = parseTimeToMinutes(formData.startTime);
-      const endMinutes = parseTimeToMinutes(formData.endTime);
-      if (startMinutes >= endMinutes) {
-        newErrors.endTime = "End time must be after start time";
-      }
-    }
-
-    if (!formData.isVirtual && !formData.location?.name?.trim()) {
-      newErrors.location = "Location name is required for offline events";
-    }
-
-    if (formData.isVirtual && !formData.virtualLink.trim()) {
-      newErrors.virtualLink = "Virtual link is required for online events";
-    }
-
-    if (formData.capacity) {
-      const capacity = Number(formData.capacity);
-      if (!capacity || capacity <= 0) {
-        newErrors.capacity = "Please enter a valid number of attendees";
-      } else if (capacity > 100000) {
-        newErrors.capacity = "Maximum capacity is 100,000 attendees";
-      }
-    }
-
-    if (formData.registrationStart || formData.registrationEnd) {
-  const now = new Date();
-
-  const registrationStart = formData.registrationStart
-    ? new Date(formData.registrationStart)
-    : null;
-
-  const registrationEnd = formData.registrationEnd
-    ? new Date(formData.registrationEnd)
-    : null;
-
-  const eventStart = new Date(
-    `${formData.isMultiDay ? formData.startDate : formData.date}T${formData.startTime}`
-  );
-
-  if (registrationStart && registrationStart < now) {
-    newErrors.registrationStart =
-      "Registration start cannot be in the past";
-  }
-
-  if (
-    registrationStart &&
-    registrationEnd &&
-    registrationStart >= registrationEnd
-  ) {
-    newErrors.registrationEnd =
-      "Registration end must be after registration start";
-  }
-
-  if (
-    registrationStart &&
-    !isNaN(eventStart.getTime()) &&
-    registrationStart >= eventStart
-  ) {
-    newErrors.registrationStart =
-      "Registration start must be before the event starts";
-  }
-
-  if (
-    registrationEnd &&
-    !isNaN(eventStart.getTime()) &&
-    registrationEnd > eventStart
-  ) {
-    newErrors.registrationEnd =
-      "Registration must close before the event starts";
-  }
-}
-
-    // Validate ticket tiers
-    if (formData.ticketTiers && formData.ticketTiers.length > 0) {
-      formData.ticketTiers.forEach((tier, index) => {
-        if (tier.name && tier.name.trim()) {
-          const price = Number(tier.price);
-          if (price < 0) {
-            newErrors[`ticketPrice_${index}`] = "Ticket price cannot be negative";
-          }
-          if (tier.capacity) {
-            const capacity = Number(tier.capacity);
-            if (capacity <= 0) {
-              newErrors[`ticketCapacity_${index}`] = "Ticket capacity must be greater than 0";
-            }
-          }
-        }
-      });
-    }
-
-    setErrors(newErrors);
-    return Object.keys(newErrors).length === 0;
-  };
 
   const handleInputChange = (e) => {
     const { name, value, type, checked } = e.target;
@@ -236,10 +100,6 @@ const EventCreation = () => {
       setErrors((prev) => ({ ...prev, [name]: "" }));
     }
   };
-
-
-
-
 
   const handleImageUpload = (e) => {
     const file = e.target.files[0];
@@ -284,29 +144,29 @@ const EventCreation = () => {
     }));
   };
 
-    const handleNext = () => {
-  try {
-    if (currentStep === CREATION_STEPS.FORM) {
-      const isValid = validateForm();
+  const handleNext = () => {
+    try {
+      if (currentStep === CREATION_STEPS.FORM) {
+        const newErrors = validateForm(formData);
+        setErrors(newErrors);
 
-      if (!isValid) {
-        toast.error("Please fix the form errors before continuing.");
-        return;
+        if (Object.keys(newErrors).length > 0) {
+          toast.error("Please fix the form errors before continuing.");
+          return;
+        }
+
+        setCurrentStep(CREATION_STEPS.PREVIEW);
+
+        window.scrollTo({
+          top: 0,
+          behavior: "smooth",
+        });
       }
-
-      setCurrentStep(CREATION_STEPS.PREVIEW);
-
-      window.scrollTo({
-        top: 0,
-        behavior: "smooth",
-      });
+    } catch (error) {
+      logger.error("Error progressing to next step:", error);
+      toast.error("Unable to continue to the next step.");
     }
-  } catch (error) {
-    logger.error("Error progressing to next step:", error);
-
-    toast.error("Unable to continue to the next step.");
-  }
-};
+  };
 
   const createEvent = () => {
     try {
@@ -394,6 +254,7 @@ const EventCreation = () => {
 
     setIsDraftLoaded(true);
   }, []);
+
   const handleRestoreDraft = () => {
     try {
       const saved = localStorage.getItem(DRAFT_KEY);
@@ -416,16 +277,14 @@ const EventCreation = () => {
 
     setShowRestoreModal(false);
   };
+
   const handleDiscardDraft = () => {
     localStorage.removeItem(DRAFT_KEY);
-
     setShowRestoreModal(false);
-
     toast.info("Saved draft discarded.");
   };
 
   useEffect(() => {
-    // Prevent saving before draft restoration
     if (!isDraftLoaded) return;
 
     const saveable = { ...formData };
@@ -435,41 +294,30 @@ const EventCreation = () => {
     localStorage.setItem(DRAFT_KEY, JSON.stringify(saveable));
   }, [formData, isDraftLoaded]);
 
-  /**
-   * Warn user before accidental refresh,
-   * tab close, or browser close
-   */
   useEffect(() => {
     const hasUnsavedChanges = Object.entries(formData).some(([key, value]) => {
-      // Ignore banner fields
       if (key === "banner" || key === "bannerPreview") {
         return false;
       }
 
-      // Handle strings
       if (typeof value === "string") {
         return value.trim() !== "";
       }
 
-      // Handle arrays
       if (Array.isArray(value)) {
         return value.length > 0;
       }
 
-      // Handle objects
       if (typeof value === "object" && value !== null) {
         return JSON.stringify(value) !== "{}";
       }
 
-      // Handle booleans/numbers
       return Boolean(value);
     });
 
     const handleBeforeUnload = (e) => {
       if (hasUnsavedChanges) {
         e.preventDefault();
-
-        // Required for browser warning
         e.returnValue = "";
       }
     };
@@ -489,113 +337,42 @@ const EventCreation = () => {
     setCurrentStep(CREATION_STEPS.FORM);
   };
 
+  const handleDurationChange = (isMultiDay) => {
+    setFormData((prev) => ({
+      ...prev,
+      isMultiDay,
+      date: "",
+      startDate: "",
+      endDate: "",
+      startTime: "",
+      endTime: "",
+    }));
+    setErrors({});
+  };
+
   return (
     <div className="min-h-screen bg-gradient-to-r from-indigo-100 to-white dark:from-gray-900 dark:to-black flex flex-col items-center justify-center py-12 px-4 sm:px-6 lg:px-8">
-      {showRestoreModal && (
-        <div
-          className="
-      fixed inset-0 z-50
-      flex items-center justify-center
-      bg-black/50
-      px-4
-    "
-        >
-          <div
-            className="
-        w-full max-w-md
-        bg-white dark:bg-gray-900
-        rounded-3xl
-        p-8
-        shadow-2xl
-        border border-gray-200
-        dark:border-gray-700
-      "
-          >
-            <h2
-              className="
-          text-2xl font-bold
-          text-gray-900 dark:text-white
-          mb-3
-        "
-            >
-              Restore Draft?
-            </h2>
-
-            <p
-              className="
-          text-gray-600 dark:text-gray-400
-          mb-6
-        "
-            >
-              A previously saved event draft was found. Would you like to restore it?
-            </p>
-
-            <div className="flex justify-end gap-3">
-              <button
-                onClick={handleDiscardDraft}
-                className="
-            px-4 py-2
-            rounded-xl
-            border border-gray-300
-            dark:border-gray-700
-            hover:bg-gray-100
-            dark:hover:bg-gray-800
-            transition
-          "
-               aria-label="button">
-                Discard
-              </button>
-
-              <button
-                onClick={handleRestoreDraft}
-                className="
-            px-5 py-2
-            rounded-xl
-            bg-indigo-600
-            hover:bg-indigo-700
-            text-white
-            font-medium
-            transition
-          "
-               aria-label="button">
-                Restore Draft
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
+      <RestoreDraftModal
+        isOpen={showRestoreModal}
+        onRestore={handleRestoreDraft}
+        onDiscard={handleDiscardDraft}
+      />
 
       {currentStep === CREATION_STEPS.FORM ? (
         <>
-          {/* Heading Section */}
           <div className="w-full max-w-4xl flex justify-end mb-6">
             <button
               onClick={() => {
                 exportAttendeesToCSV(mockAttendees, "event-attendees.csv");
-
                 toast.success("CSV exported successfully!");
               }}
-              className="
-      inline-flex
-      items-center
-      gap-2
-      px-5
-      py-3
-      rounded-2xl
-      bg-emerald-600
-      hover:bg-emerald-700
-      text-white
-      font-semibold
-      shadow-md
-      hover:shadow-lg
-      transition-all
-      duration-300
-    "
+              className="inline-flex items-center gap-2 px-5 py-3 rounded-2xl bg-emerald-600 hover:bg-emerald-700 text-white font-semibold shadow-md hover:shadow-lg transition-all duration-300"
             >
               <Download size={18} />
               Download CSV
             </button>
           </div>
+
           <motion.div
             initial={{ opacity: 0, y: -30 }}
             animate={{ opacity: 1, y: 0 }}
@@ -610,57 +387,8 @@ const EventCreation = () => {
             </p>
           </motion.div>
 
-          {/* Guidelines Section */}
-          <motion.div
-            initial={{ opacity: 0, y: 40 }}
-            whileInView={{ opacity: 1, y: 0 }}
-            viewport={{ once: true }}
-            transition={{ duration: prefersReducedMotion ? 0 : 0.7 }}
-            className="w-full max-w-4xl bg-white dark:bg-gray-800 border border-gray-200 dark:border-gray-700 shadow-lg rounded-2xl p-6 mb-10"
-          >
-            <div className="flex items-center gap-2 mb-3">
-              <ClipboardDocumentListIcon className="w-6 h-6 text-indigo-600 dark:text-indigo-400" />
-              <h2 className="text-xl font-semibold text-indigo-700 dark:text-indigo-400">
-                Guidelines
-              </h2>
-            </div>
-            <ul className="list-disc pl-6 space-y-3 text-gray-700 dark:text-gray-300 text-sm sm:text-base">
-              <li>
-                Provide a <span className="font-medium">clear and catchy title</span> that
-                accurately represents your event (3-200 characters).
-              </li>
-              <li>
-                Write a <span className="font-medium">detailed description</span> explaining what
-                attendees can expect and why they should join.
-              </li>
-              <li>
-                Set <span className="font-medium">accurate dates and times</span> to avoid
-                confusion. Make sure the end time is after the start time.
-              </li>
-              <li>
-                Choose between <span className="font-medium">virtual or in-person</span> format and
-                provide the necessary details (link or location).
-              </li>
-              <li>
-                Define <span className="font-medium">ticket tiers</span> if applicable, with clear
-                pricing and capacity limits.
-              </li>
-              <li>
-                Add relevant <span className="font-medium">tags and categories</span> to help people
-                discover your event.
-              </li>
-              <li>
-                Upload an <span className="font-medium">eye-catching banner image</span> (max 5MB)
-                to make your event stand out.
-              </li>
-              <li>
-                Review all details in the <span className="font-medium">preview</span> before
-                publishing your event.
-              </li>
-            </ul>
-          </motion.div>
+          <GuidelinesSection prefersReducedMotion={prefersReducedMotion} />
 
-          {/* Form Section */}
           <motion.div
             initial={{ opacity: 0, y: 50 }}
             animate={{ opacity: 1, y: 0 }}
@@ -668,7 +396,6 @@ const EventCreation = () => {
             className="w-full max-w-4xl bg-white dark:bg-gray-800 shadow-xl rounded-2xl p-8 border border-indigo-300 dark:border-gray-700"
           >
             <div className="space-y-6">
-              {/* General Info Step Component */}
               <GeneralInfoStep
                 formData={formData}
                 setFormData={setFormData}
@@ -680,357 +407,37 @@ const EventCreation = () => {
                 categories={categories}
               />
 
-              {/* Event Duration Type */}
-              <motion.div
-                initial={{ opacity: 0, x: -20 }}
-                whileInView={{ opacity: 1, x: 0 }}
-                viewport={{ once: true }}
-                transition={{ duration: 0.5 }}
-              >
-                <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                  <Calendar className="w-5 h-5 text-indigo-500 inline-block mr-2" />
-                  Event Duration
-                </label>
-                <div className="flex gap-6">
-                  {/* Single-day Event Option */}
-                  <label className="flex items-center text-gray-700 dark:text-white gap-2">
-                    <input
-                      type="radio"
-                      name="eventType"
-                      checked={!formData.isMultiDay}
-                      onChange={() => {
-                        setFormData((prev) => ({
-                          ...prev,
-                          isMultiDay: false,
-                          startDate: "",
-                          endDate: "",
-                          date: "",
-                          startTime: "",
-                          endTime: "",
-                        }));
-                        setErrors({});
-                      }}
-                    />
-                    Single-day Event
-                  </label>
+              <EventDurationSelector
+                isMultiDay={formData.isMultiDay}
+                onChange={handleDurationChange}
+              />
 
-                  {/* Multi-day Event Option */}
-                  <label className="flex items-center text-gray-700 dark:text-white gap-2">
-                    <input
-                      type="radio"
-                      name="eventType"
-                      checked={formData.isMultiDay}
-                      onChange={() => {
-                        setFormData((prev) => ({
-                          ...prev,
-                          isMultiDay: true,
-                          date: "",
-                          startDate: "",
-                          endDate: "",
-                          startTime: "",
-                          endTime: "",
-                        }));
-                        setErrors({});
-                      }}
-                    />
-                    Multi-day Event
-                  </label>
-                </div>
-              </motion.div>
+              <DateTimeFields
+                formData={formData}
+                handleInputChange={handleInputChange}
+                errors={errors}
+                prefersReducedMotion={prefersReducedMotion}
+                todayString={todayString}
+              />
 
-              {/* Date and Time Fields */}
-              {formData.isMultiDay ? (
-                // 🔹 Multi-day Event
-                <motion.div
-                  className="grid grid-cols-1 sm:grid-cols-4 gap-4"
-                  initial={{ opacity: 0, x: -20 }}
-                  whileInView={{ opacity: 1, x: 0 }}
-                  viewport={{ once: true }}
-                  transition={{ duration: prefersReducedMotion ? 0 : 0.5, delay: prefersReducedMotion ? 0 : 0.1 }}
-                >
-                  {/* Start Date */}
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                      Start Date <span className="text-red-600">*</span>
-                    </label>
-                    <input
-                      type="date"
-                      name="startDate"
-                      value={formData.startDate}
-                      onChange={handleInputChange}
-                      min={todayString}
-                      className={`w-full border ${
-                        errors.startDate ? "border-red-500" : "border-gray-300 dark:border-gray-600"
-                      } rounded-lg p-3 text-gray-700 dark:text-white bg-white dark:bg-gray-700`}
-                    />
-                    {errors.startDate && (
-                      <span className="text-red-500 text-sm mt-1 block">{errors.startDate}</span>
-                    )}
-                  </div>
+              <label className="flex items-center gap-2 text-sm font-medium text-gray-700 dark:text-gray-300">
+                <input
+                  type="checkbox"
+                  name="isVirtual"
+                  checked={formData.isVirtual}
+                  onChange={handleInputChange}
+                  className="w-4 h-4 text-indigo-600 border-gray-300 rounded focus:ring-indigo-500"
+                />
+                This is a virtual event
+              </label>
 
-                  {/* End Date */}
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                      End Date <span className="text-red-600">*</span>
-                    </label>
-                    <input
-                      type="date"
-                      name="endDate"
-                      value={formData.endDate}
-                      onChange={handleInputChange}
-                      min={formData.startDate || todayString}
-                      className={`w-full border ${
-                        errors.endDate ? "border-red-500" : "border-gray-300 dark:border-gray-600"
-                      } rounded-lg p-3 text-gray-700 dark:text-white bg-white dark:bg-gray-700`}
-                    />
-                    {errors.endDate && (
-                      <span className="text-red-500 text-sm mt-1 block">{errors.endDate}</span>
-                    )}
-                  </div>
+              <LocationFields
+                formData={formData}
+                handleInputChange={handleInputChange}
+                errors={errors}
+                prefersReducedMotion={prefersReducedMotion}
+              />
 
-                  {/* Start Time */}
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                      Start Time <span className="text-red-600">*</span>
-                    </label>
-                    <input
-                      type="time"
-                      name="startTime"
-                      value={formData.startTime}
-                      onChange={handleInputChange}
-                      className={`w-full border ${
-                        errors.startTime ? "border-red-500" : "border-gray-300 dark:border-gray-600"
-                      } rounded-lg p-3 text-gray-700 dark:text-white bg-white dark:bg-gray-700`}
-                    />
-                    {errors.startTime && (
-                      <span className="text-red-500 text-sm mt-1 block">{errors.startTime}</span>
-                    )}
-                  </div>
-
-                  {/* End Time */}
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                      End Time <span className="text-red-600">*</span>
-                    </label>
-                    <input
-                      type="time"
-                      name="endTime"
-                      value={formData.endTime}
-                      onChange={handleInputChange}
-                      className={`w-full border ${
-                        errors.endTime ? "border-red-500" : "border-gray-300 dark:border-gray-600"
-                      } rounded-lg p-3 text-gray-700 dark:text-white bg-white dark:bg-gray-700`}
-                    />
-                    {errors.endTime && (
-                      <span className="text-red-500 text-sm mt-1 block">{errors.endTime}</span>
-                    )}
-                  </div>
-                </motion.div>
-              ) : (
-                // 🔸 Single-day Event
-                <motion.div
-                  className="grid grid-cols-1 sm:grid-cols-3 gap-4"
-                  initial={{ opacity: 0, x: -20 }}
-                  whileInView={{ opacity: 1, x: 0 }}
-                  viewport={{ once: true }}
-                  transition={{ duration: prefersReducedMotion ? 0 : 0.5, delay: prefersReducedMotion ? 0 : 0.1 }}
-                >
-                  {/* Event Date */}
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                      Event Date <span className="text-red-600">*</span>
-                    </label>
-                    <input
-                      type="date"
-                      name="date"
-                      value={formData.date}
-                      onChange={handleInputChange}
-                      min={todayString}
-                      className={`w-full border ${
-                        errors.date ? "border-red-500" : "border-gray-300 dark:border-gray-600"
-                      } rounded-lg p-3 text-gray-700 dark:text-white bg-white dark:bg-gray-700`}
-                    />
-                    {errors.date && (
-                      <span className="text-red-500 text-sm mt-1 block">{errors.date}</span>
-                    )}
-                  </div>
-
-                  {/* Start Time */}
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                      Start Time <span className="text-red-600">*</span>
-                    </label>
-                    <input
-                      type="time"
-                      name="startTime"
-                      value={formData.startTime}
-                      onChange={handleInputChange}
-                      className={`w-full border ${
-                        errors.startTime ? "border-red-500" : "border-gray-300 dark:border-gray-600"
-                      } rounded-lg p-3 text-gray-700 dark:text-white bg-white dark:bg-gray-700`}
-                    />
-                    {errors.startTime && (
-                      <span className="text-red-500 text-sm mt-1 block">{errors.startTime}</span>
-                    )}
-                  </div>
-
-                  {/* End Time */}
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                      End Time <span className="text-red-600">*</span>
-                    </label>
-                    <input
-                      type="time"
-                      name="endTime"
-                      value={formData.endTime}
-                      onChange={handleInputChange}
-                      className={`w-full border ${
-                        errors.endTime ? "border-red-500" : "border-gray-300 dark:border-gray-600"
-                      } rounded-lg p-3 text-gray-700 dark:text-white bg-white dark:bg-gray-700`}
-                    />
-                    {errors.endTime && (
-                      <span className="text-red-500 text-sm mt-1 block">{errors.endTime}</span>
-                    )}
-                  </div>
-                </motion.div>
-              )}
-
-              {/* Virtual Event Checkbox */}
-              <motion.div
-                initial={{ opacity: 0, x: -20 }}
-                whileInView={{ opacity: 1, x: 0 }}
-                viewport={{ once: true }}
-                transition={{ duration: 0.5, delay: 0.5 }}
-              >
-                <label className="flex items-center gap-2 text-sm font-medium text-gray-700 dark:text-gray-300">
-                  <input
-                    type="checkbox"
-                    name="isVirtual"
-                    checked={formData.isVirtual}
-                    onChange={handleInputChange}
-                    className="w-4 h-4 text-indigo-600 border-gray-300 rounded focus:ring-indigo-500"
-                  />
-                  <Globe className="w-5 h-5 text-indigo-500 inline-block" />
-                  This is a virtual event
-                </label>
-              </motion.div>
-
-              {/* Virtual Link or Location */}
-              {formData.isVirtual ? (
-                <motion.div
-                  initial={{ opacity: 0, x: -20 }}
-                  whileInView={{ opacity: 1, x: 0 }}
-                  viewport={{ once: true }}
-                  transition={{ duration: 0.5 }}
-                >
-                  <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                    <Link2 className="w-5 h-5 text-indigo-500 inline-block mr-2" />
-                    Virtual Event Link <span className="text-red-600">*</span>
-                  </label>
-                  <input
-                    type="url"
-                    name="virtualLink"
-                    value={formData.virtualLink}
-                    onChange={handleInputChange}
-                    placeholder="https://zoom.us/j/..."
-                    className={`w-full border ${
-                      errors.virtualLink ? "border-red-500" : "border-gray-300 dark:border-gray-600"
-                    } rounded-lg p-3 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-1 focus:ring-indigo-500 dark:focus:ring-indigo-400 transition-all duration-300`}
-                  />
-                  {errors.virtualLink && (
-                    <span className="text-red-500 text-sm mt-1">{errors.virtualLink}</span>
-                  )}
-                </motion.div>
-              ) : (
-                <>
-                  <motion.div
-                    initial={{ opacity: 0, x: -20 }}
-                    whileInView={{ opacity: 1, x: 0 }}
-                    viewport={{ once: true }}
-                    transition={{ duration: 0.5 }}
-                  >
-                    <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                      <MapPin className="w-5 h-5 text-indigo-500 inline-block mr-2" />
-                      Location Name <span className="text-red-600">*</span>
-                    </label>
-                    <input
-                      type="text"
-                      name="location.name"
-                      value={formData.location.name}
-                      onChange={handleInputChange}
-                      placeholder="Convention Center, Community Hall, etc."
-                      className={`w-full border ${
-                        errors.location ? "border-red-500" : "border-gray-300 dark:border-gray-600"
-                      } rounded-lg p-3 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-1 focus:ring-indigo-500 dark:focus:ring-indigo-400 transition-all duration-300`}
-                    />
-                    {errors.location && (
-                      <span className="text-red-500 text-sm mt-1">{errors.location}</span>
-                    )}
-                  </motion.div>
-
-                  <motion.div
-                    initial={{ opacity: 0, x: -20 }}
-                    whileInView={{ opacity: 1, x: 0 }}
-                    viewport={{ once: true }}
-                    transition={{ duration: prefersReducedMotion ? 0 : 0.5, delay: prefersReducedMotion ? 0 : 0.1 }}
-                  >
-                    <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                      <Map className="w-5 h-5 text-indigo-500 inline-block mr-2" />
-                      Address
-                    </label>
-                    <input
-                      type="text"
-                      name="location.address"
-                      value={formData.location.address}
-                      onChange={handleInputChange}
-                      placeholder="123 Main St, City, State ZIP"
-                      className="w-full border border-gray-300 dark:border-gray-600 rounded-lg p-3 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-1 focus:ring-indigo-500 dark:focus:ring-indigo-400 transition-all duration-300"
-                    />
-                  </motion.div>
-
-                  <motion.div
-                    className="grid grid-cols-1 sm:grid-cols-2 gap-4"
-                    initial={{ opacity: 0, x: -20 }}
-                    whileInView={{ opacity: 1, x: 0 }}
-                    viewport={{ once: true }}
-                    transition={{ duration: 0.5, delay: 0.2 }}
-                  >
-                    <div>
-                      <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                        <Navigation className="w-5 h-5 text-indigo-500 inline-block mr-2" />
-                        Latitude (optional)
-                      </label>
-                      <input
-                        type="number"
-                        name="location.coordinates.latitude"
-                        value={formData.location.coordinates.latitude}
-                        onChange={handleInputChange}
-                        placeholder="40.7128"
-                        step="any"
-                        className="w-full border border-gray-300 dark:border-gray-600 rounded-lg p-3 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-1 focus:ring-indigo-500 dark:focus:ring-indigo-400 transition-all duration-300"
-                      />
-                    </div>
-
-                    <div>
-                      <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                        <Compass className="w-5 h-5 text-indigo-500 inline-block mr-2" />
-                        Longitude (optional)
-                      </label>
-                      <input
-                        type="number"
-                        name="location.coordinates.longitude"
-                        value={formData.location.coordinates.longitude}
-                        onChange={handleInputChange}
-                        placeholder="-74.0060"
-                        step="any"
-                        className="w-full border border-gray-300 dark:border-gray-600 rounded-lg p-3 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-1 focus:ring-indigo-500 dark:focus:ring-indigo-400 transition-all duration-300"
-                      />
-                    </div>
-                  </motion.div>
-                </>
-              )}
-
-              {/* Capacity */}
               <motion.div
                 initial={{ opacity: 0, x: -20 }}
                 whileInView={{ opacity: 1, x: 0 }}
@@ -1038,7 +445,6 @@ const EventCreation = () => {
                 transition={{ duration: 0.5, delay: 0.6 }}
               >
                 <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                  <Users className="w-5 h-5 text-indigo-500 inline-block mr-2" />
                   Maximum Attendees
                 </label>
                 <input
@@ -1049,67 +455,17 @@ const EventCreation = () => {
                   placeholder="Leave empty for unlimited (max: 100,000)"
                   min="1"
                   max="100000"
-                  className={`w-full border ${
-                    errors.capacity ? "border-red-500" : "border-gray-300 dark:border-gray-600"
-                  } rounded-lg p-3 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-1 focus:ring-indigo-500 dark:focus:ring-indigo-400 transition-all duration-300`}
+                  className={`w-full border ${errors.capacity ? "border-red-500" : "border-gray-300 dark:border-gray-600"} rounded-lg p-3 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-1 focus:ring-indigo-500 dark:focus:ring-indigo-400 transition-all duration-300`}
                 />
-                {errors.capacity && (
-                  <span className="text-red-500 text-sm mt-1">{errors.capacity}</span>
-                )}
+                {errors.capacity && <span className="text-red-500 text-sm mt-1">{errors.capacity}</span>}
               </motion.div>
 
-              {/* Registration Dates */}
-              <motion.div
-                className="grid grid-cols-1 sm:grid-cols-2 gap-4"
-                initial={{ opacity: 0, x: -20 }}
-                whileInView={{ opacity: 1, x: 0 }}
-                viewport={{ once: true }}
-                transition={{ duration: 0.5, delay: 0.7 }}
-              >
-                <div>
-                  <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                    <CalendarPlus className="w-5 h-5 text-indigo-500 inline-block mr-2" />
-                    Registration Start
-                  </label>
-                  <input
-                    type="datetime-local"
-                    name="registrationStart"
-                    value={formData.registrationStart}
-                    onChange={handleInputChange}
-                    className={`w-full border ${
-                      errors.registrationStart
-                        ? "border-red-500"
-                        : "border-gray-300 dark:border-gray-600"
-                    } rounded-lg p-3 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-1 focus:ring-indigo-500 dark:focus:ring-indigo-400 transition-all duration-300`}
-                  />
-                  {errors.registrationStart && (
-                    <span className="text-red-500 text-sm mt-1">{errors.registrationStart}</span>
-                  )}
-                </div>
+              <RegistrationDatesFields
+                formData={formData}
+                handleInputChange={handleInputChange}
+                errors={errors}
+              />
 
-                <div>
-                  <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                    <CalendarX className="w-5 h-5 text-indigo-500 inline-block mr-2" />
-                    Registration End
-                  </label>
-                  <input
-                    type="datetime-local"
-                    name="registrationEnd"
-                    value={formData.registrationEnd}
-                    onChange={handleInputChange}
-                    className={`w-full border ${
-                      errors.registrationEnd
-                        ? "border-red-500"
-                        : "border-gray-300 dark:border-gray-600"
-                    } rounded-lg p-3 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-1 focus:ring-indigo-500 dark:focus:ring-indigo-400 transition-all duration-300`}
-                  />
-                  {errors.registrationEnd && (
-                    <span className="text-red-500 text-sm mt-1">{errors.registrationEnd}</span>
-                  )}
-                </div>
-              </motion.div>
-
-              {/* Public and Approval Checkboxes */}
               <motion.div
                 className="space-y-3"
                 initial={{ opacity: 0, x: -20 }}
@@ -1127,7 +483,6 @@ const EventCreation = () => {
                   />
                   Make this event public
                 </label>
-
                 <label className="flex items-center gap-2 text-sm font-medium text-gray-700 dark:text-gray-300">
                   <input
                     type="checkbox"
@@ -1147,73 +502,14 @@ const EventCreation = () => {
                 setErrors={setErrors}
               />
 
-              {/* Tags Section */}
-              <motion.div
-                initial={{ opacity: 0, x: -20 }}
-                whileInView={{ opacity: 1, x: 0 }}
-                viewport={{ once: true }}
-                transition={{ duration: 0.5, delay: 1.0 }}
-              >
-                <div className="flex items-center gap-2 mb-2">
-                  <TagIcon className="w-5 h-5 text-indigo-600 dark:text-indigo-400" />
-                  <label className="text-sm font-medium text-gray-700 dark:text-gray-300">
-                    Tags
-                  </label>
-                </div>
-                <div className="flex gap-2 mb-3">
-                  <input
-                    type="text"
-                    value={newTag}
-                    onChange={(e) => setNewTag(e.target.value)}
-                    placeholder="Add a tag"
-                    onKeyDown={(e) => {
-                      if (e.key === "Enter") {
-                        e.preventDefault();
-                        addTag();
-                      }
-                    }}
-                    className="flex-1 text-sm border border-gray-300 dark:border-gray-600 rounded-lg p-2 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-1 focus:ring-indigo-500"
-                  />
-                  <button
-                    type="button"
-                    onClick={addTag}
-                    className="
-        flex items-center justify-center gap-2
-        px-4 py-2
-        rounded-3xl font-semibold
-        text-white
-        bg-black
-        shadow-md hover:shadow-lg
-        hover:bg-zinc-800
-        transform hover:scale-[1.03] active:scale-[0.97]
-        transition-all duration-300
-        focus:ring-2 focus:ring-offset-2 focus:ring-indigo-400 text-sm
-      "
-                   aria-label="button">
-                    <Plus className="w-4 h-4" />
-                    Add
-                  </button>
-                </div>
-                <div className="flex flex-wrap gap-2">
-                  {formData.tags.map((tag, index) => (
-                    <span
-                      key={index}
-                      className="inline-flex items-center gap-1 bg-indigo-100 dark:bg-indigo-900 text-indigo-700 dark:text-indigo-300 px-3 py-1 rounded-full text-sm font-medium"
-                    >
-                      #{tag}
-                      <button
-                        type="button"
-                        onClick={() => removeTag(tag)}
-                        className="ml-1 text-indigo-600 dark:text-indigo-400 hover:text-indigo-800 dark:hover:text-indigo-200 font-bold"
-                      >
-                        ×
-                      </button>
-                    </span>
-                  ))}
-                </div>
-              </motion.div>
+              <TagsInput
+                tags={formData.tags}
+                newTag={newTag}
+                onNewTagChange={setNewTag}
+                onAdd={addTag}
+                onRemove={removeTag}
+              />
 
-              {/* Submit Button */}
               <motion.button
                 type="button"
                 onClick={handleNext}
@@ -1221,37 +517,12 @@ const EventCreation = () => {
                 whileTap={{ scale: 0.98 }}
                 className="w-full flex items-center justify-center gap-2 bg-black text-white font-semibold p-4 rounded-xl shadow-lg hover:bg-zinc-800 transition-all duration-300"
               >
-                Preview Event <ArrowRightIcon className="w-5 h-5" />
+                Preview Event
               </motion.button>
             </div>
           </motion.div>
 
-          {/* Stats Section */}
-          <motion.div
-            initial={{ opacity: 0, y: 40 }}
-            whileInView={{ opacity: 1, y: 0 }}
-            viewport={{ once: true }}
-            transition={{ duration: 0.7 }}
-            className="grid grid-cols-1 sm:grid-cols-3 gap-6 w-full max-w-4xl mt-12"
-          >
-            {[
-              { number: "10k+", label: "Events Created", icon: CalendarIcon },
-              { number: "500k+", label: "Attendees", icon: UsersIcon },
-              { number: "98%", label: "Success Rate", icon: CheckCircleIcon },
-            ].map((stat, index) => (
-              <motion.div
-                key={index}
-                whileHover={{ scale: 1.08, rotate: 1 }}
-                className="bg-white dark:bg-gray-800 border border-indigo-200 dark:border-gray-700 rounded-2xl shadow-md p-6 text-center flex flex-col items-center"
-              >
-                <stat.icon className="w-10 h-10 text-indigo-600 dark:text-indigo-400 mb-3 animate-bounce" />
-                <h3 className="text-3xl font-bold text-indigo-700 dark:text-indigo-400">
-                  {stat.number}
-                </h3>
-                <p className="text-gray-600 dark:text-gray-400 mt-2">{stat.label}</p>
-              </motion.div>
-            ))}
-          </motion.div>
+          <StatsSection />
         </>
       ) : (
         <PreviewStep
